+++ rpc/shared/responses.go
@@ -10,28 +10,6 @@ import (
 // "responses.go is the interface for the types of responses for the API.
 
 /*
-"JSONResponse" is a metadata and data response in JSON format.
- */
-type JSONResponse struct {
-	Data string `json:"data"`
-}
-
-/*
-"JSONErrorResponse" is an error response in JSON format.
- */
-type JSONErrorResponse struct {
-	Error *APIError `json:"error"`
-}
-
-/*
-"APIError" is an error feedback structure containing a title and a status.
- */
-type APIError struct {
-	Status int    `json:"error"`
-	Title  string `json:"title"`
-}
-
-/*
 "WriteResponse" writes a normal JSON response.
  */
 func WriteResponse(w http.ResponseWriter, m string) {
